feat: 타일맵 콜라이더
타일맵 충돌 처리 클릭 판정 추가, 변경
#22

+++ Assets/Folder_LSR/Skills/MonsterSkillData.cs
@@ -23,7 +23,7 @@ public class MonsterSkillData : SkillBase
     public override string SkillName => skillName;
     public override ESkillType SkillType => type;
     public override ETier SkillTier => tier;
-    public override int Damage => damage;
+    public override float Damage => damage;
     public override int Coefficient => coefficient;
     public override int Range => range;
     public override int Cooldown => cooldown;
